fixed location validation issues preventing user from saving changes to va
